css-ui-3_dev/xhtml1print/box-sizing-027.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Basic User Interface Test: box-sizing:border-box used values</title>
<link href="http://florian.rivoal.net/" rel="author" title="Florian Rivoal" />
<link href="https://drafts.csswg.org/css-ui-3/#box-sizing" rel="help" />
<meta charset="utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#test {
	box-sizing: border-box;
	border: 10px solid blue;
	padding: 10px;
	width: 100px;
	height: 100px;
}
</style>
</head><body>
  <div id="test"></div>
  <div id="log"></div>

<script>
test(
  function(){
   var test = document.getElementById("test");
   assert_equals(window.getComputedStyle(test)["width"], "100px");
  }, "Check the used value of 'width' when box-sizing is  border-box.");
test(
  function(){
   var test = document.getElementById("test");
   assert_equals(window.getComputedStyle(test)["height"], "100px");
  }, "Check the used value of 'height' when box-sizing is  border-box.");
</script>


</body></html>
